requirements.txt
datarobot==2.28.0
# datarobot-bp-workshop==0.2.3
python-dotenv==0.20.0
fast-bert==2.0.1
pytorch-lamb==1.0.0
scikit-learn==1.1.1
vaderSentiment==3.3.2
scipy==1.8.1
pandas==1.3.5
spacy==3.3.0
sklearn==0.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
textblob==0.17.1
flair==0.11.3
requests==2.27.1
pip==22.1.1
beautifulsoup4==4.11.1
bs4==0.0.1
requests_futures==1.0.0
